Wrap RSS logo into <li> element
The sidebar is a <ul> element, and the RSS logo was not wrapped inside a <li> element (but a <div>). This breaks the (X)HTML standard and doesn't pass the validator.
The patch adds a <li> element while trying to keep the looking unchanged. Another way to fix it could be changing <div> to <li>.